Defer module placement on Canvas until configuration loaded - ensures ports are rendered correctly
